Split send and wallet managers
